Adding pull to refresh on songs list
