+++ app/redux/actions/songs/DispatchActions.js
@@ -38,11 +38,8 @@ export const DeleteSong = (songId, stateSongs) => {
 export const FetchSongs = (songs, songsState) => {
     if (songs.length === 0) return songsState
 
-    // If the song is already in state, don't set again
-    const newSongs = songs.filter(song => !songsState.byId[song.id])
-
     // TODO: Think if we want to use reduce vs forEach (O(n^2) vs O(n))
-    const reducedSongs = newSongs.reduce(
+    const reducedSongs = songs.reduce(
         (state, currentSongs) => {
             return {
                 byId: {
@@ -62,9 +59,8 @@ export const FetchSongs = (songs, songsState) => {
     // estrangedHD - "Well what a same, I still need another one Kappa" (02/05/20)
     reducedSongs.byId = {
         ...reducedSongs.byId,
-        ...songsState.byId,
     }
-    reducedSongs.allIds = [...reducedSongs.allIds, ...songsState.allIds]
+    reducedSongs.allIds = [...reducedSongs.allIds]
 
     return reducedSongs
 }
